% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rELAN_extraction.R
\name{extract_annotations}
\alias{extract_annotations}
\title{Extract annotations from an ELAN XML file}
\usage{
extract_annotations(file_path, wide_format = FALSE)
}
\arguments{
\item{file_path}{The file path of the ELAN XML file.}

\item{wide_format}{A boolean indicating whether to return the data frame in wide format (default: FALSE).}
}
\value{
A data frame containing the extracted annotations with adjusted time slots.
}
\description{
This function extracts annotations from an ELAN XML file and returns a data frame with adjusted time slots.
}
\examples{
annotations_df <- extract_annotations("path/to/elan/file.eaf", wide_format = FALSE)

}
