% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELANtoR.R
\name{extract_annotations}
\alias{extract_annotations}
\title{Extract Annotations from an ELAN XML Document}
\usage{
extract_annotations(elan_xml)
}
\arguments{
\item{elan_xml}{An XML document object representing the ELAN file.}
}
\value{
A tibble where each row represents an annotation.
}
\description{
Extract Annotations from an ELAN XML Document
}
\examples{
elan_xml <- xml2::read_xml(elan_xml_string)
annotations_df <- extract_annotations(elan_xml)
}
